LayoutTests/http/tests/security/cross-origin-css-5-expected.txt

CONSOLE MESSAGE: Did not parse stylesheet at 'http://127.0.0.1:8000/resources/redirect.php?url=http://localhost:8000/security/resources/xorigincss4.html' because its MIME type was invalid.

PASS Testing cross-origin and MIME behavior for CSS. 
PASS xorigincss4.html should not be loaded via @import. 
